requirements.txt
pandas
requests
bs4
PyPDF2
openpyxl
nltk
pymystem3
matplotlib
numpy
mysql.connector